Split pomerol_ed.cpp into parts
